fix: address deferred MCP review items 11-16
- server.ts: read version from package.json via createRequire instead of
  hardcoding '1.0.0'; applies to both McpServer constructor and ping tool response
- sfSpawn.ts: update SfNotFoundError to accept optional sfPath with richer
  diagnostic message (includes explicit path or 'common npm/nvm/volta locations')
- automationTools.ts: remove duplicate SfNotFoundError class; import from sfSpawn.ts
- mcp-smoke.cjs: replace shell:true with shell:process.platform==='win32' to
  avoid unnecessary shell invocation on Linux/macOS CI
- qualityHubTools.ts: replace broad /fail/i stdout scan with JSON-parse of
  result.status; falls back to word-boundary /\bFAILED?\b/i on parse failure
- licenseValidator.ts: add warning log when IDE's lastOnlineCheckMs is > 7 days
  old so stale activations are surfaced before causing auth issues
- package.json: remove 'yarn' from devDependencies (yarn is a package manager,
  not an npm package dep; GitHub Actions runners provide it via system install)
