+++ package.json
@@ -22,7 +22,6 @@
   },
   "devDependencies": {
     "@oclif/plugin-command-snapshot": "^5.0.2",
-    "yarn": "^1.22.22",
     "@salesforce/cli-plugins-testkit": "^5.1.7",
     "@salesforce/dev-scripts": "^8.3.0",
     "@types/sinon": "^21.0.0",
