+++ src/mcp/tools/qualityHubTools.ts
@@ -143,7 +143,13 @@ export function registerQualityHubTestRunReport(server: McpServer): void {
           return { isError: true as const, content: [{ type: 'text' as const, text: JSON.stringify(makeError('QH_REPORT_FAILED', result.stderr || result.stdout, requestId)) }] };
         }
 
-        const hasFailures = /fail/i.test(result.stdout);
+        let hasFailures = false;
+        try {
+          const parsed = JSON.parse(result.stdout) as { result?: { status?: string } };
+          hasFailures = /fail/i.test(parsed.result?.status ?? '');
+        } catch {
+          hasFailures = /\bFAILED?\b/i.test(result.stdout);
+        }
         const suggestion = hasFailures
           ? 'Failures detected. Use provar.qualityhub.defect.create with run_id and target_org to automatically create Defect__c records for each failure (syncs to Jira/ADO if configured).'
           : '';
